# Avoid boxing floats when calling classify_float