The PR uses the new mechanism to specify float arguments to be passed in unboxed form when calling a C function in order to optimize calls to `Pervasives.classify_float`.

A new test ensures that boxing is avoiding.

On my machine, using the mingw 32-bit port (which forces to use the emulation of classify_float, cf floats.c), a loop such as:

````ocaml
let () =
  let x = ref 1. in
  for i = 1 to 1000000000 do
    assert (classify_float !x = FP_normal);
    x := !x +. 1.
  done
````

is optimized as follow:

Before:  5.6s
After: 3.6s

And on linux 64 (in a vm):

Before: 8.7s
After: 6.6s