@mshinwell 2015-10-28
This looks good, but I think there should be a better way of making sure that the thing allocates zero.  (In Core_kernel, we have a separate C primitive for retrieving minor_words, although we always assume the result fits in the native OCaml integer width.)