@alainfrisch 2015-10-28
@mshinwell You mean, a better way to write the test, or some compiler support to guarantee that no allocation occur in a given piece of code (an attribute that triggers a warning)?