@jhjourdan 2015-10-28
This is maybe a naive remark, but wouldn't it be faster and safer to write this primitive in OCaml using the bitwise representation of floats and Int64.bits_of_float ?