@alainfrisch 2015-10-28
Another remark: using the fpclassify emulation in floats.c, even on Linux, brings even better performance:

Trunk: 8.7s
Branch: 6.6s
Branch, forcing to use the emulation: 4.0s

Does anyone know why fpclassify is slower than our emulation?