@alainfrisch 2015-10-28
@jhjourdan I doubt this could be made as fast, or even just avoid the allocation (even on 32-bit).