@mshinwell 2015-10-28
@alainfrisch For the moment, just a better test, I would have thought.  Although some kind of "no allocation allowed" attribute would be interesting.