@alainfrisch 2015-10-28
> just a better test

@mshinwell   Note that this PR only reuses the technique currently used in the testsuite, in several places, to detect the absence of allocation.