@alainfrisch 2015-10-28
@bobot I've no big plan for other functions, I had just been bitten in practice by the bad performance of classify_float and wanted to address it.  I don't see any drawback in providing unboxed variants of other functions when this is possible.