@alainfrisch 2015-10-28
While not strictly related to this discussion, it would be interesting if others could confirm that using our emulation instead of fpclassify (cf floats.c) beats the libc version (and it would be even better if someone had an explanation!).