@diml 2015-10-28
> In Core_kernel, we have a separate C primitive for retrieving minor_words, although we always assume the result fits in the native OCaml integer width.

@mshinwell, that's no problem, now we can do:

```ocaml
external minor_words : unit -> (int64 [@unboxed]) = "blah" "blah_unboxed"
```