@jhjourdan 2015-10-28
@alainfrisch, The real probelm is that Int64.bits_of_float uses a foreign C call, which seems awkward, given that it is really just identity and that most operations on Int64 are compiled directly to assembly (maybe optimizing should be a good idea).

Using a hack in order to avoid this foreign call (works only on 64bits):

```
type hack = { pad: float; mutable x :float }
let data = { pad = 0.; x = 0. }

let fpclass f =
  data.x <- f;
  let i = Obj.magic data in (* Simulate boxing of an Int64 *)
  let i = Int64.add i i in  (* Mask sign bit *)
  match Int64.shift_right i 53 with
  | 0L -> if i = 0L then FP_zero else FP_subnormal
  | -1L -> if i = 0xFFE0000000000000L then FP_infinite else FP_nan
  | _ -> FP_normal

let () =
  let x = ref 1. in
  for i = 1 to 1000000000 do
    assert (fpclass !x = FP_normal);
    x := !x +. 1.
  done
```

This gives really good performances (3.49s. vs. 7.98s. in the current trunk).