@jhjourdan 2015-10-28
And when forcing inlining of fpclass, this goes down to 1.6s.