@alainfrisch 2015-11-02
@jhjourdan This is very interesting!  Do you think good performances could be achieved on 32-bit as well?  I assume that the OCaml version is significantly slower in bytecode, which might be a bit problematic.  

Anyway, it seems the C version would still be a bit faster, even compared to the inlined OCaml version on 64-bit.