@jhjourdan 2015-11-02
@alainfrisch 64bits computations in 32bits end up in library calls, so that it will not be competitive, and yes, the performance in bytecode is bad.

Apart from that, I still think a optimized version of bits_of_float is worth it, given that it should be trivial to implement.