@xavierleroy 2015-11-02
@jhjourdan : ocamlopt native support for bits_of_float/float_of_bits involves a move between a FP register and an integer register.  This is supported by some processors (e.g. x86-SSE2, ARM) but for others (x86-x87, PowerPC) a transfer through memory is required.  Plus, this means pushing two new C-- operators down every ocamlopt back-end.  Hardly worth the effort, I think.