@alainfrisch 2015-11-02
@xavierleroy Do you have an idea why the emulation is faster than the libc version?  Could it be explained only by the avoided function call?