@xavierleroy 2015-11-02
@alainfrisch : the cost of calling fpclassify (it's not an inline function in Glibc), yes, plus the cost of converting fpclassify's return codes to the corresponding OCaml constant constructors.  

Note that if we make the emulation the default, a 64-bit version can be written that would be even faster on 64-bit platforms.