@alainfrisch 2015-11-02
@xavierleroy   That would be great and even increase the gain compared to the 64-bit version in OCaml.  So it seems we are going to stick to an external C primitive for Pervasives.classify_float.  Is anyone against merging the current PR?