@xavierleroy 2015-11-15
Commit [trunk e7f339e] improves the performance of classify_float, as discussed here.